Backed-out some unnecessary ulong -> uint32_t conversions.

+++ src/hash/crc16.c
@@ -8,7 +8,7 @@
  * @format.tab-size 4		(Plain Text/Source Code File Header)			*
  * @format.use-tabs true	(see http://www.synchro.net/ptsc_hdr.html)		*
  *																			*
- * Copyright 2003 Rob Swindell - http://www.synchro.net/copyright.html		*
+ * Copyright 2007 Rob Swindell - http://www.synchro.net/copyright.html		*
  *																			*
  * This program is free software; you can redistribute it and/or			*
  * modify it under the terms of the GNU General Public License				*
@@ -76,7 +76,7 @@ uint16_t crc16tbl[] = {
 uint16_t crc16(char* data, unsigned long len)
 {
 	uint16_t crc = 0;
-	uint32_t l;
+	unsigned long l;
 
 	if(len==0 && data!=NULL)
 		len=strlen(data);
